+++ src/wrap/result/mod.rs
@@ -1,10 +1,10 @@
 //! Result wrapper types for specialized [`Debug`](core::fmt::Debug) formatting.
 
-mod opaque_result;
-mod type_name_result;
+mod opaque;
+mod type_name;
 
 const STR_ERR: &str = "Err";
 const STR_OK: &str = "Ok";
 
-pub use opaque_result::OpaqueResult;
-pub use type_name_result::TypeNameResult;
+pub use opaque::{OpaqueResult, OpaqueResultMarker};
+pub use type_name::TypeNameResult;