+++ src/wrap/type_name.rs
@@ -20,8 +20,8 @@ use crate::types::{DisplayMode, Full, Short};
 ///
 /// # Type Parameters
 ///
-/// - `T`: The type who's name to display. Same as `V` for value wrappers.
-/// - `V`: The value to wrap. [`PhantomData<fn() -> T>`](PhantomData) for marker types.
+/// - `D`: The type who's name to display. Same as `V` for value wrappers.
+/// - `T`: The value to wrap. [`PhantomData<fn() -> T>`](PhantomData) for marker types.
 /// - `M`: The display mode.
 ///
 /// # Examples
@@ -39,20 +39,20 @@ use crate::types::{DisplayMode, Full, Short};
 /// assert_eq!(*wrapped, vec![1, 2, 3]); // Can still access the value
 /// ```
 #[derive(Copy, Clone, Deref, AsMut, AsRef, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct TypeName<T: ?Sized = (), V = PhantomData<fn() -> T>, M: DisplayMode = Short>(
+pub struct TypeName<D: ?Sized = (), T = PhantomData<fn() -> D>, M: DisplayMode = Short>(
     /// The wrapped value, if any. For marker types, this is [`PhantomData<fn() -> T>`](PhantomData).
     #[deref]
     #[as_mut]
     #[as_ref]
-    pub V,
-    pub(crate) PhantomData<(fn() -> T, M)>,
+    pub T,
+    pub(crate) PhantomData<(fn() -> D, M)>,
 );
 
 #[allow(clippy::mismatching_type_param_order, reason = "T is used for both Value and Type")]
 impl<T> TypeName<T, T, Full> {
-    /// Wrap a value, displaying its type name in debug output.
+    /// Wrap a value in a [`TypeName`], displaying its type name in debug output.
     ///
-    /// The type is inferred from the value.
+    /// The display type is inferred from the value.
     ///
     /// # Examples
     ///
@@ -61,7 +61,6 @@ impl<T> TypeName<T, T, Full> {
     ///
     /// let wrapped = TypeName::wrap::<Short>(42i32);
     /// assert_eq!(format!("{:?}", wrapped), "i32");
-    /// assert_eq!(*wrapped, 42, "Wrapped value is accessible");
     ///
     /// let wrapped = TypeName::wrap::<Full>(vec![1]);
     /// assert_eq!(format!("{:?}", wrapped), "alloc::vec::Vec<i32>");
@@ -69,9 +68,29 @@ impl<T> TypeName<T, T, Full> {
     pub fn wrap<M: DisplayMode>(value: T) -> TypeName<T, T, M> {
         TypeName(value, PhantomData)
     }
+
+    /// Wrap a borrowed value in a [`TypeName`], displaying its type name in debug output.
+    ///
+    /// The type is inferred from the value.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use display_as_debug::types::{TypeName, Full, Short};
+    ///
+    /// let vec = vec![1];
+    /// let wrapped = TypeName::borrow::<Short>(&vec);
+    /// assert_eq!(format!("{:?}", wrapped), "Vec<i32>");
+    ///
+    /// let wrapped = TypeName::borrow::<Full>(&vec);
+    /// assert_eq!(format!("{:?}", wrapped), "alloc::vec::Vec<i32>");
+    /// ```
+    pub fn borrow<M: DisplayMode>(value: &T) -> TypeName<T, &T, M> {
+        TypeName(value, PhantomData)
+    }
 }
 
-impl<T, V, M: DisplayMode> TypeName<T, V, M> {
+impl<D, T, M: DisplayMode> TypeName<D, T, M> {
     /// Consumes the wrapper, returning the inner value.
     ///
     /// # Examples
@@ -82,20 +101,20 @@ impl<T, V, M: DisplayMode> TypeName<T, V, M> {
     /// let wrapped = TypeName::wrap::<Short>(42i32);
     /// assert_eq!(wrapped.into_inner(), 42);
     /// ```
-    pub fn into_inner(self) -> V {
+    pub fn into_inner(self) -> T {
         self.0
     }
 }
 
-#[allow(clippy::mismatching_type_param_order, reason = "T is used for both Value and Type")]
+#[allow(clippy::mismatching_type_param_order, reason = "T is used for both Display and Value")]
 impl<T, M: DisplayMode> From<T> for TypeName<T, T, M> {
     fn from(value: T) -> Self {
         Self(value, PhantomData)
     }
 }
 
-impl<T: ?Sized, V, M: DisplayMode> Debug for TypeName<T, V, M> {
+impl<D: ?Sized, T, M: DisplayMode> Debug for TypeName<D, T, M> {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        f.write_str(M::type_name::<T>())
+        f.write_str(M::type_name::<D>())
     }
 }