+++ tests/types/opaque/mod.rs
@@ -3,4 +3,6 @@
 mod list;
 #[allow(clippy::module_inception, reason = "Matches implementation module name")]
 mod opaque;
+mod option;
+mod result;
 mod set;