+++ tests/types/type_name/type_name.rs
@@ -10,9 +10,6 @@ mod full {
 
     test_fmt!(consts, TypeName::<Vec<i32>>::FULL, "{:?}", EXPECTED);
     test_fmt!(empty, TypeName::empty::<Vec<i32>, Full>(), "{:?}", EXPECTED);
-    test_fmt!(wrap, TypeName::wrap::<Full>(vec![1]), "{:?}", EXPECTED);
-    test_fmt!(from, TypeName::<i32, i32, Full>::from(42), "{:?}", "i32");
-    test_get!(into_inner, TypeName::wrap::<Full>(42i32), move TypeName::into_inner, 42);
 }
 
 mod short {
@@ -22,7 +19,4 @@ mod short {
 
     test_fmt!(consts, TypeName::<Vec<i32>>::SHORT, "{:?}", EXPECTED);
     test_fmt!(empty, TypeName::empty::<Vec<i32>, Short>(), "{:?}", EXPECTED);
-    test_fmt!(wrap, TypeName::wrap::<Short>(vec![1]), "{:?}", EXPECTED);
-    test_fmt!(from, TypeName::<i32, i32, Short>::from(42), "{:?}", "i32");
-    test_get!(into_inner, TypeName::wrap::<Short>(42i32), move TypeName::into_inner, 42);
 }